src/test/kotlin/me/peckb/aoc/_2017/calendar/day05/Day05Test.kt
package me.peckb.aoc._2017.calendar.day05

import javax.inject.Inject


import me.peckb.aoc._2017.DaggerTestDayComponent
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.BeforeEach
import org.junit.jupiter.api.Test

internal class Day05Test {
  @Inject
  lateinit var day05: Day05

  @BeforeEach
  fun setup() {
    DaggerTestDayComponent.create().inject(this)
  }

  @Test
  fun testDay05PartOne() {
    assertEquals(342669, day05.partOne(DAY_05))
  }

  @Test
  fun testDay05PartTwo() {
    assertEquals(25136209, day05.partTwo(DAY_05))
  }

  companion object {
    private const val DAY_05: String = "src/test/resources/2017/day05.input"
  }
}
